validation completed

+++ Power_Usage_App-1/src/main/java/com/power/model/PowerRange.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class PowerRange {
 	
 	int day;
-	int unitconsumed;
+	double unitconsumed;
 	
 
 }
